requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.6
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-compressor==4.4
django-icons==24.3
django-libsass==0.9
django-phonenumber-field==7.3.0
django-sass==1.1.0
django-sass-processor==1.4
fontawesomefree==6.4.2
libsass==0.23.0
rcssmin==1.1.1
rjsmin==1.2.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3